Handle uri local paths and refactor into helper class
